backend/pyproject.toml
[project]
name = "churnvision-backend"
version = "0.1.0"
description = "ChurnVision Enterprise Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "email-validator>=2.0.0",
    "sqlalchemy>=2.0.25",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.3",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-text-splitters>=0.0.1",
    "pandas>=2.2.0",
    "httpx>=0.26.0",
    "openai>=1.10.0",
    "langchain-openai>=0.0.5",
    "ollama>=0.1.0",
    "PyJWT>=2.8.0",
    "greenlet>=3.0.3",
    "bcrypt>=4.0.1",
    # RAG subsystem dependencies
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.5.0",
    # Survival analysis
    "lifelines>=0.27.0",
    # ML explainability and validation
    "shap>=0.44.0",
    "imbalanced-learn>=0.11.0",
    # Monitoring
    "prometheus-fastapi-instrumentator>=6.1.0",
    # SSO/OIDC
    "authlib>=1.3.0",
    "itsdangerous>=2.1.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Caching
    "redis>=5.0.0",
    # Encryption
    "cryptography>=41.0.0",
    # Structured logging
    "structlog>=24.1.0",
    # Background tasks
    "celery[redis]>=5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "httpx>=0.26.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
